+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## Unreleased
 
-- compile routes.
+- compile routes in production.
 - (set! *default-data-reader-fn* tagged-literal)
-- konpy.db wrapper?
-- java.lang.NullPointerException when empty week-num or task provided.
+- self answers
+- ohter answers
+- tailwindcss
+
+
+## 0.6.7-SNAPSHOT
 
 ## 0.6.6 - hotfix